Move exportFastA functions into shared utils instead of inheritable abstract view
