+++ public/scripts/plugins/sequencing_primers/views/sequencing_primers_products_view.js
@@ -1,8 +1,14 @@
-import AbstractViewContainingSequences from '../../common/views/abstract_view_containing_sequences';
+import {fastAExportSequenceFromID} from '../../../common/lib/utils';
 import template from '../templates/sequencing_primers_products_view.hbs';
 
-export default AbstractViewContainingSequences.extend({
+
+export default Backbone.View.extend({
+  manage: true,
   template: template,
+
+  events: {
+    'click .export-sequence': 'exportSequence',
+  },
 
   serialize: function() {
     return {
@@ -14,6 +20,16 @@ export default AbstractViewContainingSequences.extend({
     this.$('.has-tooltip').tooltip({
       container: 'body'
     });
+  },
+
+  getProducts: function() {
+    return this.parentView().products;
+  },
+
+  exportSequence: function(event) {
+    var sequenceID = $(event.target).data('sequence_id');
+    var products = this.getProducts();
+    fastAExportSequenceFromID(products, sequenceID);
   },
 
 });
